refactor: separated the page state, view and helper widget logics into different files.
